Layth claim crud (#12)
* change application.properties configuration

* switch to main

+++ mobilityBackend/src/main/java/com/gurus/mobility/entity/alert/Alert.java
@@ -21,7 +21,7 @@ public class Alert implements Serializable {
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public int idAlrt;
+    public Long idAlrt;
     @NotNull
     public String alrtName;
     @Enumerated(EnumType.STRING)
